2011-12-30-14.15.24: src/c/machine-dependent/posix-arithmetic-trap-handlers.c: Modified arithmetic_fault_handler() to call dump_all() before die()ing.

+++ src/c/heapcleaner/heap-debug-stuff.c
@@ -216,7 +216,7 @@ static void   dump_gen0_tripwire_buffers__guts   (FILE* fd, Task* task, char* ca
 	//
 	for (int i = 0;  i < AGEGROUP0_OVERRUN_TRIPWIRE_BUFFER_SIZE_IN_WORDS;  ++i) {
 	    //
-	    fprintf(fd,"%p: %08x", p+i, p[i]);
+	    fprintf(fd,"%p: %08x", p+i, (unsigned int) p[i]);
 	    if (p[i] != 0) 	        	fprintf(fd,"   # Overrun!");
 	    fprintf(fd,"\n");
 	}
